+++ src/main/java/hu/jakab/ekkeencoprosampbackend/dto/measurementUnit/MeasurementUnitCreatedDTO.java
@@ -2,6 +2,7 @@ package hu.jakab.ekkeencoprosampbackend.dto.measurementUnit;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Data
@@ -10,7 +11,7 @@ public class MeasurementUnitCreatedDTO {
     private String unitCode;
     private String description;
     private String unitCategory;
-    private Double conversionFactor;
+    private BigDecimal conversionFactor;
     private String standardBody;
 
 }
